+++ CHANGELOG.md
@@ -1,52 +1,71 @@
+## 2.0.12
+
+- fixed loading issue.
+- manages selected permission.
+- safe area issue.
+
 ## 2.0.11
+
 - create custom theme for bottom sheet.
 - Camera UI changes.
 - plugin upgrade.
 
 ## 2.0.5
+
 - added web support for photo,documents,audio,video from storage.
 
 ## 2.0.4
- - Added audio pick functionality through a bottom sheet for Android.
- - Implemented audio file picking for both Android and iOS.
+
+- Added audio pick functionality through a bottom sheet for Android.
+- Implemented audio file picking for both Android and iOS.
 
 ## 2.0.3
+
 - Performance improvement.
 - Image preview , video info
 
 ## 2.0.2
+
 - Ui changes
 
 ## 2.0.1
+
 - bug fix in ios choose Document
 
-
 ## 1.0.10
+
 - added document picker
 
-
 ## 1.0.8
+
 - release mode not working bug fixed
 
 ## 1.0.7
+
 - allow video recording from camera
 
 ## 1.0.6
+
 - bug fix
 
 ## 1.0.5
+
 - Added zoom controls in camera
 
 ## 1.0.4
+
 - Added capture image from camera
 
 ## 1.0.3
+
 - bug fix
 
 ## 1.0.2
+
 - bug fix
 
 ## 1.0.1
+
 - Initial release of `picker_pro_max_ultra`.
 - Features:
     - Image and video picking from gallery.